web socket connection instead of post so that game is being updated for everyone
